code/requirements.txt
qiskit
pennylane
networkx
numpy
scipy
matplotlib
web3
gql
requests
python-dotenv
